requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
Django==5.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
